+++ src/app/categories/[slug]/page.tsx
@@ -3,15 +3,11 @@ import { notFound } from 'next/navigation';
 import { Col, Container, Row } from 'react-bootstrap';
 
 import { BookID } from '@data/contrarianmba.json';
-import { BookCategoryView } from '@/components/Book';
 import { CategoryTitle } from '@/components/CategoryTitle';
+import { CategoryBooks } from '@/components/CategoryBooks';
 import { Book } from '@/types';
 import { CATEGORIES } from '@/constants';
-import {
-    lookUpBookById,
-    lookUpBookIdsByCategory,
-    resolveCategorySlug,
-} from '@/utils';
+import { resolveCategorySlug } from '@/utils';
 
 type Props = {
     params: {
@@ -20,26 +16,11 @@ type Props = {
 };
 export default function CategoryPage({ params: { slug } }: Props) {
     const category = resolveCategorySlug(slug);
-    const bookIds: string[] = lookUpBookIdsByCategory(category);
-    if (typeof bookIds === 'undefined') {
-        notFound();
-    }
-    const books: Book[] = bookIds.map((bookId: BookID) =>
-        lookUpBookById(bookId)
-    );
 
     return (
         <>
             <CategoryTitle category={category} />
-            <Container>
-                <Row>
-                    {books.map((book: Book) => (
-                        <Col lg={3} key={book.id}>
-                            <BookCategoryView book={book} />
-                        </Col>
-                    ))}
-                </Row>
-            </Container>
+            <CategoryBooks category={category} />
         </>
     );
 }
