+++ src/app/page.tsx
@@ -1,24 +1,30 @@
 'use client';
-import { Hero } from '@/components/Hero';
-import data from '@data/contrarianmba.json';
+
 import { Col, Container, Row } from 'react-bootstrap';
-import { BookCategoryView } from '../components/Book';
+
+import { CATEGORIES } from '@/constants';
+import { Hero } from '@/components/Hero';
+import { CategoryTitle } from '@/components/CategoryTitle';
+import { CategoryBooks } from '@/components/CategoryBooks';
 
 export default function Home() {
-    const books = Object.values(data.lookups.book_id);
-
     return (
         <>
             <Hero />
-            <Container>
-                <Row>
-                    {books.map((book) => (
-                        <Col key={book.id} lg={3}>
-                            <BookCategoryView book={book} />
-                        </Col>
-                    ))}
-                </Row>
-            </Container>
+            {CATEGORIES.map((category) => {
+                return (
+                    <p>
+                        <Container>
+                            <Row>
+                                <Col>
+                                    <h3>{category.name}</h3>
+                                </Col>
+                            </Row>
+                        </Container>
+                        <CategoryBooks category={category} />
+                    </p>
+                );
+            })}
         </>
     );
 }
